fix unit test error
